requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
python-dotenv
